Add update on change feature and threshold setting (#4)
* Add update on change feature and threshold setting

This commit adds the "UpdateOnChange" feature and "UpdateOnChangeThreshold" setting to the application. The "UpdateOnChange" feature allows the application to update only when there is a change in the input image. The "UpdateOnChangeThreshold" setting determines the percentage of change required to trigger an update. These new features improve the efficiency of the application by reducing unnecessary updates.

* lint

* Remove unnecessary log statements and add error handling in tesseract_thread function

* Refactor tesseract-ocr-utils.cpp

* Add Output Formatting option to en-US.ini